bmt/requirements.txt
asgiref==3.8.1
Django==5.1.7
sqlparse==0.5.3
tzdata==2025.2
django-unfold==0.54.0
django-import-export==4.3.7
djangorestframework==3.16.0
celery==5.5.1
redis==5.2.1
